requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
pandas>=2.1.10
openpyxl==3.1.2
Pillow==10.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
